feat: documents listing with dynamic categories + market breadcrumbs + BreadcrumbList JSON-LD
- Documents listing sidebar now loads categories dynamically from backend
- Added Home / Market / <Page> breadcrumbs to Prices, Lots, Buyers, Farmers, Documents, Bookings
- Added BreadcrumbList JSON-LD structured data to all six market sub-pages
- Extracted BookingsClient component so bookings page.tsx can be a server component
