+++ apps/client-fnp/app/(landing)/documents/page.tsx
@@ -7,6 +7,16 @@ export const metadata = {
     description: "Download premium farm management guides, spray programs, agrochemical datasheets and seed catalogues.",
 }
 
+const breadcrumbJsonLd = {
+    "@context": "https://schema.org",
+    "@type": "BreadcrumbList",
+    "itemListElement": [
+        { "@type": "ListItem", "position": 1, "name": "Home", "item": "https://farmnport.com" },
+        { "@type": "ListItem", "position": 2, "name": "Market", "item": "https://farmnport.com/market" },
+        { "@type": "ListItem", "position": 3, "name": "Documents", "item": "https://farmnport.com/documents" },
+    ],
+}
+
 export default async function DocumentsPage() {
     let initialDocs: any[] = []
     let initialTotal = 0
@@ -21,9 +31,12 @@ export default async function DocumentsPage() {
 
     return (
         <div className="min-h-screen bg-gradient-to-b from-background to-muted/20">
+            <script type="application/ld+json" dangerouslySetInnerHTML={{ __html: JSON.stringify(breadcrumbJsonLd) }} />
             <div className="mx-auto max-w-7xl px-6 lg:px-8 py-6">
                 <nav className="flex items-center gap-1.5 text-xs text-muted-foreground mb-6">
                     <Link href="/" className="hover:text-foreground transition-colors">Home</Link>
+                    <span>/</span>
+                    <Link href="/market" className="hover:text-foreground transition-colors">Market</Link>
                     <span>/</span>
                     <span className="text-foreground font-medium">Documents</span>
                 </nav>
