+++ apps/client-fnp/app/(landing)/prices/page.tsx
@@ -1,3 +1,4 @@
+import Link from "next/link"
 import { PricesBoard } from "@/components/structures/prices-board"
 
 export const metadata = {
@@ -14,9 +15,28 @@ export const metadata = {
 }
 
 export default function PricesPage() {
+  const breadcrumbJsonLd = {
+    "@context": "https://schema.org",
+    "@type": "BreadcrumbList",
+    "itemListElement": [
+      { "@type": "ListItem", "position": 1, "name": "Home", "item": "https://farmnport.com" },
+      { "@type": "ListItem", "position": 2, "name": "Market", "item": "https://farmnport.com/market" },
+      { "@type": "ListItem", "position": 3, "name": "Prices", "item": "https://farmnport.com/prices" },
+    ],
+  }
   return (
     <>
+      <script type="application/ld+json" dangerouslySetInnerHTML={{ __html: JSON.stringify(breadcrumbJsonLd) }} />
       <h1 className="sr-only">Agricultural Market Prices Zimbabwe – Livestock, Produce &amp; More</h1>
+      <div className="mx-auto max-w-7xl px-6 lg:px-8 pt-4">
+        <nav className="flex items-center gap-1.5 text-xs text-muted-foreground mb-4">
+          <Link href="/" className="hover:text-foreground transition-colors">Home</Link>
+          <span>/</span>
+          <Link href="/market" className="hover:text-foreground transition-colors">Market</Link>
+          <span>/</span>
+          <span className="text-foreground font-medium">Prices</span>
+        </nav>
+      </div>
       <PricesBoard />
     </>
   )
